project/requirements.txt
django
google-cloud-aiplatform
google-generativeai
python-dotenv
google-cloud-discoveryengine
